Update gitIgnore Steam Commands now send user info message in DM. Problems with deleting old message

+++ Modules/MessageCommands.cs
@@ -1,13 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-using System.Threading;
-
+﻿using Discord;
 using Discord.Commands;
-using Discord;
+using System;
+using System.Linq;
+using System.Threading.Tasks;
 
 namespace UsefulDiscordBot.Modules
 {
@@ -16,7 +11,7 @@ namespace UsefulDiscordBot.Modules
 		[Group("clear")]
     public class MessageCommands : ModuleBase<SocketCommandContext>
     {
-				bool isAdmin => (Context.User as IGuildUser).Guild.Roles.FirstOrDefault(x => x.Name == "Mods") != null;
+				bool isAdmin => (Context.User as IGuildUser)?.Guild?.Roles.FirstOrDefault(x => x.Name == "Mods") != null;
 
 				[Command("from")]
         public async Task deleteFrom([Remainder] string user = "")
@@ -42,52 +37,65 @@ namespace UsefulDiscordBot.Modules
 								}
 								await ReplyAsync($"`{Context.User.Username} deleted {Amount} messages`");
 						}
-						await ReplyAsync("You must be a moderator to delete messages");
-        }
+						else
+						{
+								await ReplyAsync("You do not have Admin privelages");
+						}
+				}
 
         [Command()]
         public async Task clear([Remainder] int Delete = 0)
         {
-            //await checkPermissions();
-            if (Delete == null)
-            {
-                await Context.Channel.SendMessageAsync("`You need to specify the amount | !clear (amount) | Replace (amount) with anything`");
-            }
-            int Amount = 0;
+						if (isAdmin)
+						{
+								//await checkPermissions();
+								if (Delete == 0)
+								{
+										await Context.Channel.SendMessageAsync("`You need to specify the amount | !clear (amount) | Replace (amount) with any number`");
+										return;
+								}
 
-            Console.WriteLine("Deleting all Messages");
-            foreach (var Item in await Context.Channel.GetMessagesAsync(Delete + 1).Flatten())
-            {
+								Console.WriteLine("Deleting {0} messages", Delete);
+								var items = await Context.Channel.GetMessagesAsync(Delete + 1).Flatten();								
+								await Context.Channel.DeleteMessagesAsync(items);
 
-                Amount++;
-                await Item.DeleteAsync();
-
-            }
-            await Context.Channel.SendMessageAsync($"`{Context.User.Username} deleted {Amount} messages`");
-        }
+								await Context.Channel.SendMessageAsync($"`{Context.User.Username} deleted {Delete} messages`");
+						}
+						else
+						{
+								await ReplyAsync("You do not have Admin privelages");
+						}
+				}
 
         [Command("containing")]
         public async Task delete([Remainder] string key = "")
         {
-            //await checkPermissions();
-            if (key == string.Empty)
-            {
-                await Context.Channel.SendMessageAsync("`You need to specify the key | !clear \"key\" | Replace \"key\" with any string");
-            }
+						if (isAdmin)
+						{
+								//await checkPermissions();
+								if (key == string.Empty)
+								{
+										await Context.Channel.SendMessageAsync("`You need to specify the key | !clear \"key\" | Replace \"key\" with any string");
+								}
 
-            int Amount = 0;
+								int Amount = 0;
 
-            Console.WriteLine("Deleting Messages");
-            foreach (var Item in await Context.Channel.GetMessagesAsync(1000).Flatten())
-            {
-                if (Item.Content.Contains(key))
-                {
-                    Amount++;
-                    await Item.DeleteAsync();
-                }
+								Console.WriteLine("Deleting Messages");
+								foreach (var Item in await Context.Channel.GetMessagesAsync(1000).Flatten())
+								{
+										if (Item.Content.Contains(key))
+										{
+												Amount++;
+												await Item.DeleteAsync();
+										}
 
-            }
-            await Context.Channel.SendMessageAsync($"`{Context.User.Username} deleted {Amount} messages`");
+								}
+								await Context.Channel.SendMessageAsync($"`{Context.User.Username} deleted {Amount} messages`");
+						}
+						else
+						{
+								await ReplyAsync("You do not have Admin privelages");
+						}
         }
 
         public async Task checkPermissions()
